requirements.txt
streamlit==1.28.0
tensorflow-cpu==2.15.0
Pillow==10.2.0
numpy==1.23.5
h5py==3.11.0
protobuf==3.20.3